PUBDEV-3797: API tests for H2O Module completed. (#693) (#751)
PUBDEV-3797: API test for H2O Module in Python.  Incorporated Pasha code reivew to use the assert_is_type in h2o.utils.typecheck instead of my own.  Deleted mine from pyunit_utils.

PUBDEV-3797: API test for H2O Module in Python.  Incorporated Pasha code reivew to remove quotes from frame id when calling h2o.api().

PUBDEV-3797: api test for PYthon commands in H2O module.  Added to shutdown test to show Pasha that error in one thread is caught correctly.

PUBDEV-3797: API test for H2O Module.  Incorporate Pasha code review comments for show_progress,no_progress and get_grid.

PUBDEV-3797:  add api tests for H2O Module in python.  Removed remove_all from tests.  It is causing key not found in other tests.

PUBDEV-3793: API tests for Python H2O Module.  Will test h2o.remove_all() by inspecting and make sure its argument list is empty.

PUBDEV-3797: api tests for PYthon H2O Module.  Change tests for h2o.import_sql_table and h2o.imoprt_sql_select.  We are checking their argument lists
