+++ h2o-py/tests/testdir_apis/H2O_Module/pyunit_h2oinit.py
@@ -1,6 +1,4 @@
 from __future__ import print_function
-from builtins import str
-from builtins import range
 import sys
 sys.path.insert(1,"../../../")
 from tests import pyunit_utils
