+++ h2o-py/tests/testdir_apis/H2O_Module/pyunit_h2ono_progress.py
@@ -8,7 +8,6 @@ try:
     from StringIO import StringIO
 except ImportError:
     from io import StringIO
-from h2o.utils.typechecks import assert_is_type
 
 def h2ono_progress():
     """
