+++ h2o-py/tests/testdir_apis/H2O_Module/pyunit_h2oremove_all.py
@@ -3,29 +3,19 @@ import sys
 sys.path.insert(1,"../../../")
 from tests import pyunit_utils
 import h2o
-from h2o.estimators.glm import H2OGeneralizedLinearEstimator
-from h2o.utils.typechecks import assert_is_type
-from pandas import DataFrame
+import inspect
 
 def h2oremove_all():
     """
     Python API test: h2o.remove_all()
+
+    Cannot test this one on Jenkins.  It will crash other tests.  So, Pasha found a way around this
+    by just checking the argument list which should be empty.
     """
     # call with no arguments
     try:
-        h2o.remove_all()    # remove all object first
-        training_data = h2o.import_file(pyunit_utils.locate("smalldata/logreg/benign.csv"))
-        Y = 3
-        X = [0, 1, 2, 4, 5, 6, 7, 8, 9, 10]
-
-        model = H2OGeneralizedLinearEstimator(family="binomial", alpha=0, Lambda=1e-5)
-        model.train(x=X, y=Y, training_frame=training_data)
-        lsObject = h2o.ls()
-        assert_is_type(lsObject, DataFrame)
-        assert len(lsObject.values) == 3, "h2o.ls() command is not working."
-        h2o.remove_all()
-        lsObject2 = h2o.ls()
-        assert len(lsObject2.values) == 0, "h2o.remove_all() command is not working."
+        allargs = inspect.getargspec(h2o.remove_all)
+        assert len(allargs.args)==0, "h2o.remove_all() should have no arguments!"
     except Exception as e:
         assert False, "h2o.remove_all() command is not working."
